requirements.txt
# Основные зависимости
paramiko>=3.0.0
pydantic>=2.0.0
pyyaml>=6.0
loguru>=0.7.0
tenacity>=8.0.0
rich>=13.0.0
typer>=0.9.0

# Дополнительные зависимости для SSH Agent
cryptography>=41.0.0
scp>=0.14.0
fabric>=3.0.0

# LLM интеграция
openai>=1.0.0
anthropic>=0.7.0

# Асинхронность
aiofiles>=23.0.0
asyncio-mqtt>=0.13.0

# HTTP клиенты
httpx>=0.25.0
aiohttp>=3.8.0
requests>=2.31.0

# Тестирование
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
pytest-cov>=4.0.0

# Веб-интерфейс (опционально)
fastapi>=0.100.0
uvicorn>=0.23.0

# Мониторинг
prometheus-client>=0.17.0

# Утилиты
python-dotenv>=1.0.0
click>=8.0.0

# Дополнительные зависимости для SSH Agent
cryptography>=41.0.0
scp>=0.14.0
fabric>=3.0.0

# Конфигурация и валидация
jsonschema>=4.0.0
marshmallow>=3.20.0

# Логирование и отладка
structlog>=23.0.0
colorama>=0.4.6

# Безопасность
keyring>=24.0.0
cryptography>=41.0.0